requirements.txt
# requirements.txt
ccxt==2.6.92
cryptography==38.0.4
pandas==1.5.3
numpy==1.24.3
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2